kvserver/rangefeed: move helper functions to registry_helpers_test
This patch moves helper functions to registry_helpers_test.

Part of: #129814
Release note: none
